Refactor structure and correct in line with mentor's comments
