+++ requirements.txt
@@ -1,0 +1,8 @@
+email-validator==1.1.3
+Flask==2.0.1
+Flask-Bootstrap==3.3.7.1
+Flask-Migrate==3.0.1
+Flask-SQLAlchemy==2.5.1
+Flask-WTF==0.15.1
+psycopg2==2.9.1
+requests==2.25.1
