+++ tests/test_final_task.py
@@ -1,6 +1,3 @@
-import pytest
-
-
 # Place for tests
 def test_1():
     assert 1 == 1
